# Conflicts:
#	src/main/java/com/green/acamatch/academy/Service/AcademyService.java

+++ src/main/java/com/green/acamatch/user/service/UserManagementService.java
@@ -9,7 +9,6 @@ import com.green.acamatch.config.security.AuthenticationFacade;
 import com.green.acamatch.user.UserUtils;
 import com.green.acamatch.user.entity.User;
 import com.green.acamatch.user.model.UserDeleteReq;
-import com.green.acamatch.user.model.UserSignInRes;
 import com.green.acamatch.user.model.UserUpdateReq;
 import com.green.acamatch.user.repository.UserRepository;
 import jakarta.servlet.http.HttpServletResponse;
@@ -26,52 +25,50 @@ import java.io.IOException;
 public class UserManagementService {
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
-    private final SignUpUserCache signUpUserCache;
+    private final UserCache userCache;
     private final UserUtils userUtils;
     private final MyFileUtils myFileUtils;
     private final UserConst userConst;
 
     @Transactional
-    public UserSignInRes signUp(String token, HttpServletResponse response) {
-        User user = signUpUserCache.verifyToken(token);
+    public void signUp(String token, HttpServletResponse response) {
+        User user = userCache.verifyToken(token);
         userRepository.save(user);
         try {
-            response.sendRedirect("/");
+            response.sendRedirect(userConst.getRedirectUrl());
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
-        return userUtils.generateUserSignInResByUser(user, response);
     }
 
-    public int updateUserPic(MultipartFile pic) {
-        long userId = AuthenticationFacade.getSignedUserId();
-        User user = userUtils.getUserById(userId);
-
-        String prePic = user.getUserPic();
-        user.setUserPic(myFileUtils.makeRandomFileName(pic));
-        userRepository.save(user);
-
-        String folderPath = String.format(userConst.getUserPicFilePath(), userId);
-        if (prePic != null) {
-            myFileUtils.deleteFolder(folderPath, false);
-        }
-        myFileUtils.makeFolders(folderPath);
-        String filePath = String.format("%s/%s", folderPath, user.getUserPic());
-        try {
-            myFileUtils.transferTo(pic, filePath);
-        } catch (IOException e) {
-            throw new CustomException(CommonErrorCode.INTERNAL_SERVER_ERROR);
-        }
-        return 1;
-    }
-
-    public int updateUser(UserUpdateReq req) {
+    @Transactional
+    public int updateUser(UserUpdateReq req, MultipartFile pic) {
         User user = userUtils.getUserById(AuthenticationFacade.getSignedUserId());
-        if(req.getName() != null) user.setName(req.getName());
-        if(req.getNickName() != null) user.setNickName(req.getNickName());
-        if(req.getBirth() != null) user.setBirth(req.getBirth());
-        if(req.getPhone() != null) user.setPhone(req.getPhone());
+        if (!passwordEncoder.matches(req.getCurrentPw(), user.getUpw())) {
+            throw new CustomException(UserErrorCode.INCORRECT_PW);
+        }
+        if (req.getName() != null) user.setName(req.getName());
+        if (req.getNickName() != null) user.setNickName(req.getNickName());
+        if (req.getBirth() != null) user.setBirth(req.getBirth());
+        if (req.getPhone() != null) user.setPhone(req.getPhone());
+        if (req.getNewPw() != null) user.setUpw(passwordEncoder.encode(req.getNewPw()));
+        if (pic != null) {
+            String prePic = user.getUserPic();
+            user.setUserPic(myFileUtils.makeRandomFileName(pic));
+            String folderPath = String.format(userConst.getUserPicFilePath(), user.getUserId());
+            if (prePic != null) {
+                myFileUtils.deleteFolder(folderPath, false);
+            }
+            myFileUtils.makeFolders(folderPath);
+            String filePath = String.format("%s/%s", folderPath, user.getUserPic());
+            try {
+                myFileUtils.transferTo(pic, filePath);
+            } catch (IOException e) {
+                throw new CustomException(CommonErrorCode.INTERNAL_SERVER_ERROR);
+            }
+        }
         userRepository.save(user);
+
         return 1;
     }
 
@@ -86,5 +83,17 @@ public class UserManagementService {
         String folderPath = String.format(userConst.getUserPicFilePath(), userId);
         myFileUtils.deleteFolder(folderPath, true);
         return 1;
+    }
+
+    public void setTempPw(long id, HttpServletResponse response) {
+        String pw = userCache.getTempPw(id);
+        User user = userUtils.getUserById(id);
+        user.setUpw(passwordEncoder.encode(pw));
+        userRepository.save(user);
+        try {
+            response.sendRedirect(userConst.getRedirectUrl());
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
     }
 }
